# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: HireSphere API
  version: 1.0.0
  description: API documentation for the HireSphere interview booking platform.
servers:
  - url: http://localhost:5000
    description: Local development server
...
paths:
...
  /bookings/candidate/{candidateId}:
    get:
      tags: [Bookings]
      summary: List bookings for a candidate
      parameters:
        - $ref: "#/components/parameters/CandidateId"
      responses:
        "200":
          description: Candidate bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
  /bookings/interviewer/{interviewerId}:
    get:
      tags: [Bookings]
      summary: List bookings for an interviewer
      parameters:
        - $ref: "#/components/parameters/InterviewerId"
      responses:
        "200":
          description: Interviewer bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
...
  /messages:
    post:
      tags: [Messages]
      summary: Create a booking message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageRequest"
      responses:
        "201":
          description: Message created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "400":
          $ref: "#/components/responses/BadRequest"
  /messages/{bookingId}:
    get:
      tags: [Messages]
      summary: List messages for a booking
      parameters:
        - $ref: "#/components/parameters/BookingId"
      responses:
        "200":
          description: Booking messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
components:
  parameters:
...
    CandidateId:
      name: candidateId
      in: path
      required: true
      schema:
        type: integer
      example: 1
    InterviewerId:
      name: interviewerId
      in: path
      required: true
      schema:
        type: integer
      example: 1
    BookingId:
      name: bookingId
      in: path
      required: true
      schema:
        type: integer
      example: 1
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: Invalid request
...
    Booking:
      type: object
      properties:
        id:
          type: integer
          example: 1
        candidate_id:
          type: integer
          example: 1
        interviewer_id:
          type: integer
          example: 1
        candidate_name:
          type: string
          example: Demo Candidate
        interviewer_name:
          type: string
          example: Demo Interviewer
        domain:
          type: string
          example: Backend
        interview_type:
          type: string
          example: System Design
        booking_date:
          type: string
          format: date-time
        status:
          type: string
          example: booked
        meeting_link:
          type: string
          example: https://meet.example.com/session
        submission_file_name:
          nullable: true
          type: string
          example: solution.zip
        evaluation_feedback:
          nullable: true
          type: string
          example: Strong technical explanation.
        evaluation_score:
          nullable: true
          type: integer
          example: 86
...
    MessageRequest:
      type: object
      required: [bookingId, senderId, receiverId, message]
      properties:
        bookingId:
          type: integer
          example: 1
        senderId:
          type: integer
          example: 1
        receiverId:
          type: integer
          example: 2
        message:
          type: string
          example: I uploaded the submission.
    Message:
      allOf:
        - $ref: "#/components/schemas/MessageRequest"
        - type: object
          properties:
            id:
              type: integer
              example: 1
            sender_name:
              type: string
              example: Demo Candidate
            receiver_name:
              type: string
              example: Demo Interviewer
            created_at:
              type: string
              format: date-time